+++ test/system_tar.dart
@@ -25,7 +25,7 @@ Future<Process> startTar(List<String> args) {
 
 Future<Process> writeToTar(List<String> args, Stream<tar.Entry> entries) async {
   final proc = await startTar(args);
-  await entries.pipe(tar.WritingSink(proc.stdin));
+  await entries.pipe(tar.createWritingSink(proc.stdin));
 
   return proc;
 }