+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.0-dev
 
 - Remove `MemoryEntry`. Use `Entry.data` to create a tar entry from bytes.
+- Make `WritingSink` private. Use `createWritingSink` to create a general `StreamSink<tar.Entry>`.
+- Make `Reader` private. Use `createReader` to create a custom tar reader.
 
 ## 0.1.0-nullsafety.1